+++ implant/Cargo.toml
@@ -6,7 +6,8 @@ edition = "2021"
 [dependencies]
 tonic = "0.9"
 prost = "0.11"
-tokio = { version = "1.0", features = ["macros", "rt-multi-thread"] }
+tokio = { version = "1.0", features = ["macros", "rt-multi-thread", "process"] }
+rand = "0.8.5"
 
 [build-dependencies]
 tonic-build = "0.9"
